+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- SimpleCheckpoint utility
+
 ## [8.6.0] - 2024-05-09
 
 ### Added